requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
pillow==11.3.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
python-dateutil==2.9.0.post0
gunicorn==22.0.0
whitenoise==6.5.0
drf-spectacular==0.27.2
openpyxl==3.1.2
